requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile
#
bidict==0.22.1
    # via python-socketio
blinker==1.7.0
    # via flask
click==8.1.7
    # via flask
colorama==0.4.6
    # via click
flask==3.0.0
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-socketio
    #   flask-sqlalchemy
flask-cors==4.0.0
    # via -r requirements.in
flask-socketio==5.3.6
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via -r requirements.in
greenlet==3.0.1
    # via sqlalchemy
h11==0.14.0
    # via wsproto
importlib-metadata==6.8.0
    # via flask
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
psycopg2==2.9.9
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pyftpdlib==1.5.9
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
python-engineio==4.8.0
    # via python-socketio
python-socketio==5.10.0
    # via flask-socketio
simple-websocket==1.0.0
    # via python-engineio
sqlalchemy==2.0.23
    # via flask-sqlalchemy
typing-extensions==4.8.0
    # via sqlalchemy
werkzeug==3.0.1
    # via flask
wsproto==1.2.0
    # via simple-websocket
zipp==3.17.0
    # via importlib-metadata
